Redefine stateEquality to compare value of names
This was done to fix an issue with stateValid. Because the state names
were string literals defined in a header file, when two modules that
include it are compiled separately, we can end up having two copies of
the string in the program with two separate addresses. This necessitates
using strCmp to check equality instead of simply comparing the addresses