Remove average values from state_counting args
This does mean recomputing what those averages are in state_counting,
but I think the code is cleaner this way because it means one fewer
way in which `argsCounting` can be invalid. However, the real reason for
this change is that a forthcoming commit will have `state_starting` pass
additional frames to `state_counting`. That commit would require some
ugly code were it not for this commit, particularly because
`state_counting` doesn't already know the average value of those frames.